refactor(pty): extract env resolution into environment.ts
Move SANITIZED_ENV_KEYS, SANITIZED_ENV_PREFIXES, getSystemPath(),
resolveSystemPath(), execToString(), and sanitizedEnv() out of the
1300-line pseudoterminal.ts into a dedicated src/terminal/environment.ts.
pseudoterminal.ts now imports sanitizedEnv() from environment.js.

Rename tests/src/terminal/system-path.spec.ts to environment.spec.ts to
mirror the new source file name.
